refactor(clientinfo): merge keepcommon/clientinfo registry into pkg/clientinfo
pkg/clientinfo previously wrapped a keepcommon/clientinfo.Registry via
embedding purely to add a ctx field and custom metric/diagnostic
registration helpers on top. Since no other project depends on
pkg/keepcommon as a library, the indirection buys nothing: merge the
base registry (metrics/diagnostics storage, gauge/info/observer
primitives) directly into pkg/clientinfo instead.
