+++ pkg/clientinfo/clientinfo.go
@@ -1,17 +1,31 @@
+// Package clientinfo provides tools for gathering and exposing system
+// metrics and diagnostics for external monitoring tools.
+//
+// Currently, this package is intended to use with Prometheus but can be
+// easily extended if needed. Also, not all Prometheus metric types are
+// implemented.
+//
+// Following specifications were used as reference:
+// - https://prometheus.io/docs/instrumenting/writing_clientlibs/
+// - https://prometheus.io/docs/instrumenting/exposition_formats/
 package clientinfo
 
 import (
 	"context"
+	"io"
 	"net/http"
 	"net/http/pprof"
+	"sort"
+	"strconv"
+	"sync"
 	"time"
 
 	"github.com/ipfs/go-log"
-
-	"github.com/keep-network/keep-core/pkg/keepcommon/clientinfo"
 )
 
 var logger = log.Logger("keep-clientinfo")
+
+const readHeaderTimeout = 2 * time.Second
 
 // Config stores configuration for the client info.
 type Config struct {
@@ -26,12 +40,27 @@ type Config struct {
 	EnablePprof bool
 }
 
-// Registry wraps keep-common clientinfo registry and exposes additional
-// functions for registering client-custom metrics and diagnostics
+// Registry performs all management of metrics and diagnostics. Specifically,
+// it allows registering and exposing them through the HTTP server, and
+// exposes additional functions for registering client-custom metrics and
+// diagnostics.
 type Registry struct {
-	*clientinfo.Registry
-
 	ctx context.Context
+
+	metrics      map[string]metric
+	metricsMutex sync.RWMutex
+
+	diagnosticsSources map[string]func() string
+	diagnosticsMutex   sync.RWMutex
+}
+
+// newRegistry creates a new client info registry bound to ctx.
+func newRegistry(ctx context.Context) *Registry {
+	return &Registry{
+		ctx:                ctx,
+		metrics:            make(map[string]metric),
+		diagnosticsSources: make(map[string]func() string),
+	}
 }
 
 // Initialize set up the client info registry and enables metrics and
@@ -44,15 +73,14 @@ func Initialize(
 		return nil, false
 	}
 
-	registry := &Registry{clientinfo.NewRegistry(), ctx}
+	registry := newRegistry(ctx)
 
 	if cfg.EnablePprof {
 		// Register the pprof handlers on http.DefaultServeMux, which is the
-		// mux that keep-common's EnableServer hands to the http.Server.
-		// Registering them explicitly here avoids the side-effecting blank
-		// import of net/http/pprof, which would otherwise register
-		// /debug/pprof/* unconditionally on DefaultServeMux regardless of
-		// this flag.
+		// mux that EnableServer hands to the http.Server. Registering them
+		// explicitly here avoids the side-effecting blank import of
+		// net/http/pprof, which would otherwise register /debug/pprof/*
+		// unconditionally on DefaultServeMux regardless of this flag.
 		registerPprofHandlers()
 		logger.Infof("pprof profiling endpoints enabled at /debug/pprof/")
 	}
@@ -72,4 +100,42 @@ func registerPprofHandlers() {
 	http.HandleFunc("/debug/pprof/profile", pprof.Profile)
 	http.HandleFunc("/debug/pprof/symbol", pprof.Symbol)
 	http.HandleFunc("/debug/pprof/trace", pprof.Trace)
+}
+
+// EnableServer enables the client info server on the given port. Data will
+// be exposed on `/metrics` and `/diagnostics` paths.
+func (r *Registry) EnableServer(port int) {
+	server := &http.Server{
+		Addr:              ":" + strconv.Itoa(port),
+		ReadHeaderTimeout: readHeaderTimeout,
+	}
+
+	http.HandleFunc("/metrics", func(response http.ResponseWriter, _ *http.Request) {
+		if _, err := io.WriteString(response, r.exposeMetrics()); err != nil {
+			logger.Errorf("could not write response: [%v]", err)
+		}
+	})
+
+	http.HandleFunc("/diagnostics", func(response http.ResponseWriter, _ *http.Request) {
+		if _, err := io.WriteString(response, r.exposeDiagnostics()); err != nil {
+			logger.Errorf("could not write response: [%v]", err)
+		}
+	})
+
+	go func() {
+		if err := server.ListenAndServe(); err != http.ErrServerClosed {
+			logger.Errorf("client info server error: [%v]", err)
+		}
+	}()
+}
+
+func sortedKeys[V any](m map[string]V) []string {
+	keys := make([]string, len(m))
+	i := 0
+	for k := range m {
+		keys[i] = k
+		i++
+	}
+	sort.Strings(keys)
+	return keys
 }
