+++ pkg/clientinfo/gauge.go
