+++ pkg/clientinfo/gauge_test.go
