+++ pkg/clientinfo/info_test.go
