+++ pkg/clientinfo/observer.go
