+++ pkg/clientinfo/registry_diagnostics.go
