+++ pkg/clientinfo/registry_metrics.go
